old sync user support

+++ renderer/src/modules/projects/Sync.js
@@ -82,7 +82,7 @@ export default function Sync() {
         // for pull without conflict
         const checkoutFIles = await checkoutJsonFiles(pullData.fs, pullData.gitprojectDir, pullData.checkoutBranch);
         const pullStatus = checkoutFIles && await pullProject(pullData.fs, pullData.gitprojectDir, pullData.userBranch, auth.token.sha1, pullData.checkoutBranch);
-        pullStatus?.status && await updateSettingsFiles(
+      pullStatus?.status && await updateSettingsFiles(
           pullData.fs,
           pullData.sbDataObject,
           pullData.projectDir,
@@ -97,36 +97,40 @@ export default function Sync() {
         await addNotification('Sync', 'Project Sync Successfull', 'success');
       } else {
         // delete project + clone the project
-        await pullData?.fs.rmdir((pullData?.gitprojectDir), { recursive: true }, async (err) => {
+        const exist = await pullData?.fs.existsSync(pullData?.gitprojectDir);
+        let clone = true;
+        exist && await pullData?.fs.rmdir((pullData?.gitprojectDir), { recursive: true }, async (err) => {
           if (err) {
             logger.debug('Sync.js', 'Error removing project directory for clone');
             // throw new Error(`Remove Resource failed :  ${err}`);
-          } else {
-            // call clone
-            const cloneStatus = await cloneAndSetProject(
-              pullData.fs,
-              pullData.gitprojectDir,
-              pullData.repo,
-              pullData.userBranch,
-              pullData.auth,
-              pullData.checkoutBranch,
-            );
-            // continue settings file writing
-            cloneStatus && await updateSettingsFiles(
-              pullData.fs,
-              pullData.sbDataObject,
-              pullData.projectDir,
-              pullData.projectName,
-              pullData.id,
-              pullData.currentUser,
-              pullData.updateBurrito,
-              pullData.action,
-            );
+            clone = false;
           }
-        logger.debug('Sync.js', 'Project Sync to scribe successfull, clone successfull');
-        await notifyStatus('success', 'Project Sync to scribe successfull');
-        await addNotification('Sync', 'Project Sync Successfull', 'success');
-      });
+        });
+        if (clone) {
+          // call clone
+          const cloneStatus = await cloneAndSetProject(
+            pullData.fs,
+            pullData.gitprojectDir,
+            pullData.repo,
+            pullData.userBranch,
+            pullData.auth,
+            pullData.checkoutBranch,
+          );
+          // continue settings file writing
+          cloneStatus && await updateSettingsFiles(
+            pullData.fs,
+            pullData.sbDataObject,
+            pullData.projectDir,
+            pullData.projectName,
+            pullData.id,
+            pullData.currentUser,
+            pullData.updateBurrito,
+            pullData.action,
+          );
+          logger.debug('Sync.js', 'Project Sync to scribe successfull, clone successfull');
+          await notifyStatus('success', 'Project Sync to scribe successfull');
+          await addNotification('Sync', 'Project Sync Successfull', 'success');
+        }
       }
     } else {
         logger.debug('Sync.js', 'error pullData not set from function');
